BTN-1631 - Make activity service non blocking. Previously, waiting for the procedure to be created meant the message republish logic was blocked, so the procedure created message wasn't received by the function which was waiting for it.
This change breaks that blocking, by subscribing at the worker level to the procedure created message. It introduces a bit more state into the service and relies on the request_ids more, which I don't love. Some of this could be refactored once we've thought a bit more about the activity states and topics